requirements.txt
Telethon>=1.36.0
cryptg>=0.5.0
python-dotenv>=1.0.1
